Implement driver option IgnoreEDID
Since some time, savage enforces video-mode and max. pixel clock
validation through EDID, which is great for autoconfiguration, but
prevents users from defining higher screen resolutions, even though the
monitor is capable of it. This patch implements the IgnoreEDID option so
that the EDID monitor information can be explictly disabled.