Savage: Do not add panel modes when there is no panel
SavageGetPanelInfo will reset DisplayType if the panel is not
active, so check again before calling SavageAddPanelMode so that
bogus modes are not added.

This fixes startup failures in some cases, for instance when there is no
real panel and EDID is missing. The bogus sync ranges (typically 0 Hz)
would also block the default sync ranges from being applied.

Thanks to Peter Belew for testing.
https://bugs.launchpad.net/bugs/294899