savage: Fix Xv scaling on non-native resolutions
If I set the resolution to 800x600 for example (to be able to use the
TV output in full screen mode), I only can see 800/1024 = 0,78 of the
real video width, no matter if the video is displayed in a small player
window or enlarged to full screen. On the right of the video display,
the remaining width is filled with a black (or sometimes blue) bar.

The problem was mainly that only the drawing starting point was modified
by the scaling factor, but not the size-defining end of the drawing
rectangle.

https://bugs.launchpad.net/bugs/670790

Signed-off-by: Hans-Juergen Mauser <hjmauser@gmx.net>
Reviewed-by: Tormod Volden <debian.tormod@gmail.com>
Acked-by: Matt Turner <mattst88@gmail.com>