savage: Avoid unnecessary float calculation in video display
There is no reason to do the division using floats as long as
the result is an integer and there is no risk of intermediate
overflow.

This is in the code for "Old" chipsets. Integer division was already
used in the parallel code for "New" chipsets.

Also fixed up the whitespace in both blocks.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>
Reviewed-by: Matt Turner <mattst88@gmail.com>